monitor: Change factor from size_t to double
To support a broader range of possible factors users could apply, change
the optional factor from a size_t to double type. This allows a
configuration to use a decimal factor value.

Tested: N/A

+++ monitor/tach_sensor.cpp
@@ -74,7 +74,7 @@ TachSensor::TachSensor(Mode mode,
                        bool hasTarget,
                        size_t funcDelay,
                        const std::string& interface,
-                       size_t factor,
+                       double factor,
                        int64_t offset,
                        size_t timeout,
                        const sdeventplus::Event& event) :
